feat(git): migrate to callback-based API
We were using a vendored version of |Vital.Async.Promise| previously,
but it didn't make much sense since users don't have access to our
version of `s:Promise`. Also, it is easy for a user to adapt the
callback-based API to their own version of `Promise` if they want,
anyway.

This paves the way for the stabilization of asynchronous trust sources
API.